[!!!][TASK] Drop social image wizard and prefer ext:seo (#121)
We are dropping the social image wizard and rely on the editor to create
custom images if necessary. The feature predated the core SEO-Initiative
that now handles social images through dedicated fields. We are strongly
recommending using core handling instead of the media fields.

The wizard itself had several drawbacks, it was hard to configure and
was not able to handle high-resolution images.